Moved command to it's own modual. This involved making coconut.command.utils and making coocnut.const
